fix: Product Blocks not being linked to Products with create()
When using the `create()` helper, Products should be linked to Product Blocks when one puts a list of names for the create function (see function docstring). This was not happening correctly because I was iterating through the wrong list when trying to resolve the names to UUIDs (old: `product_blocks`, now: `products`). This PR fixes that problem.