Replace update_ticket_step with addtional_steps
The optional update_ticket_step argument of the create_workflow,
modify_workflow and terminate_workflow decorators is implemented in a
non-optional way. This is resolved by replacing it with an additional_steps
argument that takes a StepList instead of a Step. Besides now being able to add
more than one additional step, as will be used by the Generate Workflows CLI
command, the step list can also be empty.

Note that this will be a breaking change for those who use update_ticket_step
in there custom orchestrator.