[#724] Exclude subscriptions_search materialized view from Alembic autogeneration (#877)
* [#724] Exclude subscriptions_search materialized view from Alembic autogeneration

The subscriptions_search view is a manually created materialized view and should not be managed by Alembic. This commit adds an info flag to the view’s table definition and updates the include_object logic to skip any table marked as a materialized view. This prevents erroneous migration generation for subscriptions_search. Resolves #724.

---------

Co-authored-by: Mohammad Torkashvand <mohammad.torkashvand@geant.org>