Improve step timestamping logic (#991)
* Improve step timestamping logic

This commit creates the new started_at field for the process_steps table and renames executed_at to completed_at.
Corresponding fields are also created in the REST and GraphQL APIs.
The executed_at field in the process_steps table and the "executed" field in the API is left in place for backwards compatibility(but marked deprecated)
Step start time is computed immediately before the step is run (in the step decorator's wrapper function), injected into the state, and later removed from the state and used to populate the DB field.
A backfill migration is done to populate started at for previously run steps either with the executed time of the previous step, or the start time of the workflow(if it's the first step)

* Fixed linting issues

* Fix migration downgrade; make started_at timestamping work across step types

* Add support for start_at for step groups

* Changes requested in code review

* Got process.map working for state updates

---------

Co-authored-by: John Gann <jgann@es.net>